chore: separate nonce plugin from html plugin
