+++ packages/core/src/plugins/html.ts
@@ -1,10 +1,8 @@
 import path, { isAbsolute } from 'node:path';
 import {
   type MinifyJSOptions,
-  applyToCompiler,
   castArray,
   color,
-  createVirtualModule,
   deepmerge,
   fse,
   getDistPath,
@@ -417,40 +415,12 @@ export const pluginHtml = (modifyTagsFn?: ModifyHTMLTagsFn): RsbuildPlugin => ({
       },
     );
 
-    api.onAfterCreateCompiler(({ compiler }) => {
-      const { nonce } = api.getNormalizedConfig().security;
-
-      if (!nonce) {
-        return;
-      }
-
-      applyToCompiler(compiler, (compiler) => {
-        const { plugins } = compiler.options;
-        const hasHTML = plugins.some(
-          (plugin) => plugin && plugin.constructor.name === 'HtmlBasicPlugin',
-        );
-        if (!hasHTML) {
-          return;
-        }
-
-        // apply __webpack_nonce__
-        // https://webpack.js.org/guides/csp/
-        const injectCode = createVirtualModule(
-          `__webpack_nonce__ = "${nonce}";`,
-        );
-        new compiler.webpack.EntryPlugin(compiler.context, injectCode, {
-          name: undefined,
-        }).apply(compiler);
-      });
-    });
-
     api.modifyHTMLTags({
       // ensure `crossorigin` and `nonce` can be applied to all tags
       order: 'post',
       handler: ({ headTags, bodyTags }) => {
         const config = api.getNormalizedConfig();
         const { crossorigin } = config.html;
-        const { nonce } = config.security;
         const allTags = [...headTags, ...bodyTags];
 
         if (crossorigin) {
@@ -463,15 +433,6 @@ export const pluginHtml = (modifyTagsFn?: ModifyHTMLTagsFn): RsbuildPlugin => ({
               (tag.tag === 'link' && tag.attrs?.rel === 'stylesheet')
             ) {
               tag.attrs.crossorigin ??= formattedCrossorigin;
-            }
-          }
-        }
-
-        if (nonce) {
-          for (const tag of allTags) {
-            if (tag.tag === 'script' || tag.tag === 'style') {
-              tag.attrs ??= {};
-              tag.attrs.nonce = nonce;
             }
           }
         }
